crawler_service/requirements.txt
psycopg[binary]
fastapi
uvicorn
playwright
python_dotenv